+++ DtoLibPos/Documento/Entidad/Ficha.cs
@@ -119,7 +119,7 @@ namespace DtoLibPos.Documento.Entidad
         public decimal MontoPorVueltoEnEfectivo { get; set; }
         public decimal MontoPorVueltoEnDivisa { get; set; }
         public decimal MontoPorVueltoEnPagoMovil { get; set; }
-        public int CantDivisaPorVueltoEnDivisa { get; set; }
+        public decimal CantDivisaPorVueltoEnDivisa { get; set; }
 
 
         public Ficha()
@@ -228,17 +228,17 @@ namespace DtoLibPos.Documento.Entidad
             MontoPorVueltoEnEfectivo = 0m;
             MontoPorVueltoEnDivisa = 0m;
             MontoPorVueltoEnPagoMovil = 0m;
-            CantDivisaPorVueltoEnDivisa = 0;
+            CantDivisaPorVueltoEnDivisa = 0m;
             //
             BonoPorPagoDivisa = 0m;
             MontoBonoPorPagoDivisa = 0m;
-            CntDivisaAplicaBonoPorPagoDivisa = 0;
+            CntDivisaAplicaBonoPorPagoDivisa = 0m;
         }
 
 
         public decimal BonoPorPagoDivisa { get; set; }
         public decimal MontoBonoPorPagoDivisa { get; set; }
-        public int CntDivisaAplicaBonoPorPagoDivisa { get; set; }
+        public decimal CntDivisaAplicaBonoPorPagoDivisa { get; set; }
     }
 
 }
